refactor(web): move mapper components to lib/mappers/components
