+++ appeals/web/src/server/appeals/appeal-details/appellant-case/address/address.mapper.js
@@ -1,5 +1,5 @@
 import { appealShortReference } from '#lib/appeals-formatter.js';
-import { addressInputs } from '#lib/page-components/address.js';
+import { addressInputs } from '#lib/mappers/components/address.js';
 
 /**
  * @typedef {import('../../appeal-details.types.js').WebAppeal} Appeal
