+++ appeals/web/src/server/appeals/appeal-details/appellant-case/application-decision-date/application-decision-date.mapper.js
@@ -2,7 +2,7 @@
  * @typedef {import('../../appeal-details.types.js').WebAppeal} Appeal
  */
 import { appealShortReference } from '#lib/appeals-formatter.js';
-import { yesNoInput } from '#lib/page-components/radio.js';
+import { yesNoInput } from '#lib/mappers/components/radio.js';
 import { dateISOStringToDayMonthYearHourMinute } from '#lib/dates.js';
 
 /**
