+++ appeals/web/src/server/appeals/appeal-details/lpa-questionnaire/correct-appeal-type/correct-appeal-type.mapper.js
@@ -1,5 +1,5 @@
 import { appealShortReference } from '#lib/appeals-formatter.js';
-import { yesNoInput } from '#lib/page-components/radio.js';
+import { yesNoInput } from '#lib/mappers/components/radio.js';
 
 /**
  * @typedef {import('../../appeal-details.types.js').WebAppeal} Appeal
