+++ appeals/web/src/server/appeals/appeal-details/lpa-questionnaire/extra-conditions/extra-conditions.mapper.js
@@ -4,7 +4,7 @@
 
 import { appealShortReference } from '#lib/appeals-formatter.js';
 import { convertFromYesNoToBoolean } from '#lib/boolean-formatter.js';
-import { yesNoInput } from '#lib/page-components/radio.js';
+import { yesNoInput } from '#lib/mappers/components/radio.js';
 
 /**
  * @param {Appeal} appealData
