+++ appeals/web/src/server/appeals/appeal-details/safety-risks/safety-risks.mapper.js
@@ -3,7 +3,7 @@
  */
 
 import { appealShortReference } from '#lib/appeals-formatter.js';
-import { yesNoInput } from '#lib/page-components/radio.js';
+import { yesNoInput } from '#lib/mappers/components/radio.js';
 
 /**
  * @param {Appeal} appealData
