+++ appeals/web/src/server/lib/mappers/appeal/submappers/appellant.mapper.js
@@ -1,5 +1,5 @@
 import { formatServiceUserAsHtmlList } from '#lib/service-user-formatter.js';
-import { textDisplayField } from '#lib/page-components/text.js';
+import { textDisplayField } from '#lib/mappers/components/text.js';
 
 /** @type {import('../appeal.mapper.js').SubMapper} */
 export const mapAppellant = ({ appealDetails, currentRoute, userHasUpdateCasePermission }) =>
